+++ TestApps/UnitTestApp/FunctionalTests/ANBannerAdViewVideoTestCase.m
@@ -39,16 +39,17 @@
     // Put teardown code here. This method is called after the invocation of each test method in the class.
     [super tearDown];
     [[ANHTTPStubbingManager sharedStubbingManager] disable];
+    [[ANHTTPStubbingManager sharedStubbingManager] removeAllStubs];
     self.banner.delegate = nil;
     self.banner.appEventDelegate = nil;
     [self.banner removeFromSuperview];
     self.banner = nil;
+    self.loadAdSuccesfulException = nil;
 }
 
 
 - (void)clearSetupBannerVideoAd {
     self.banner = nil;
-    [[ANHTTPStubbingManager sharedStubbingManager] removeAllStubs];
 }
 
 -(void) setupBannerVideoAd{