+++ sdk/internal/ANGlobal.h
@@ -27,7 +27,7 @@
 #define AN_ERROR_TABLE @"errors"
 
 #define AN_DEFAULT_PLACEMENT_ID		@"default_placement_id"
-#define AN_SDK_VERSION                  @"5.3"
+#define AN_SDK_VERSION                  @"5.4"
 
 
 #define APPNEXUS_BANNER_SIZE		CGSizeMake(320, 50)
@@ -127,6 +127,9 @@ NSURLRequest *ANBasicRequestWithURL(NSURL *URL);
 NSNumber *ANiTunesIDForURL(NSURL *URL);
 BOOL ANCanPresentFromViewController(UIViewController *viewController);
 
+
+static NSString *const kANCreativeId= @"creativeId";
+static NSString *const kANImpressionUrls = @"impressionUrls";
 
 
 #pragma mark - Global class.