+++ sdk/internal/ANUniversalAdFetcher.m
@@ -32,13 +32,13 @@
 #import "ANTrackerInfo.h"
 #import "ANTrackerManager.h"
 #import "NSTimer+ANCategory.h"
+#import "ANNativeRenderingViewController.h"
+#import "ANRTBNativeAdResponse.h"
 
-
-
-
-@interface ANUniversalAdFetcher () <ANVideoAdProcessorDelegate, ANAdWebViewControllerLoadingDelegate, ANNativeMediationAdControllerDelegate>
+@interface ANUniversalAdFetcher () <ANVideoAdProcessorDelegate, ANAdWebViewControllerLoadingDelegate, ANNativeMediationAdControllerDelegate , ANNativeRenderingViewControllerLoadingDelegate>
 
 @property (nonatomic, readwrite, strong)  ANMRAIDContainerView              *adView;
+@property (nonatomic, readwrite, strong)  ANNativeRenderingViewController       *nativeAdView;
 @property (nonatomic, readwrite, strong)  ANMediationAdViewController       *mediationController;
 @property (nonatomic, readwrite, strong)  ANNativeMediatedAdController      *nativeMediationController;
 @property (nonatomic, readwrite, strong)  ANSSMMediationAdViewController    *ssmMediationController;
@@ -184,7 +184,7 @@
         [self handleSSMMediatedAd:nextAd];
         
     } else if ( [nextAd isKindOfClass:[ANNativeStandardAdResponse class]] ) {
-        [self handleNativeStandardAd:nextAd];
+        [self handleNativeAd:nextAd];
         
     } else {
         ANLogError(@"Implementation error: Unknown ad in ads waterfall.  (class=%@)", [nextAd class]);
@@ -344,13 +344,67 @@
                                                                   adViewDelegate:self.delegate];
 }
 
-- (void)handleNativeStandardAd:(ANNativeStandardAdResponse *)nativeStandardAd
+- (void)handleNativeAd:(ANNativeStandardAdResponse *)nativeAd
 {
     
-    ANAdFetcherResponse  *fetcherResponse  = [ANAdFetcherResponse responseWithAdObject:nativeStandardAd andAdObjectHandler:nil];
-    [self processFinalResponse:fetcherResponse];
+
+    BOOL enableNativeRendering = NO;
+    if ([self.delegate respondsToSelector:@selector(enableNativeRendering)]) {
+        enableNativeRendering = [self.delegate enableNativeRendering];
+        if (([nativeAd.nativeRenderingUrl length] > 0) && enableNativeRendering){
+            ANRTBNativeAdResponse *rtnNativeAdResponse = [[ANRTBNativeAdResponse alloc] init];
+            rtnNativeAdResponse.nativeAdResponse = nativeAd ;
+            [self renderNativeAd:rtnNativeAdResponse];
+            return;
+        }
+    }
+    // Traditional native ad instance.
+    [self traditionalNativeAd:nativeAd];
+ 
 }
 
+-(void)traditionalNativeAd:(ANNativeStandardAdResponse *)nativeAd{
+    ANAdFetcherResponse  *fetcherResponse = [ANAdFetcherResponse responseWithAdObject:nativeAd andAdObjectHandler:nil];
+    [self processFinalResponse:fetcherResponse];
+
+}
+
+-(void) renderNativeAd:(ANBaseAdObject *)nativeRenderingElement {
+    
+    CGSize sizeofWebView = [self getAdSizeFromDelegate];
+    
+    
+    if (self.nativeAdView) {
+        self.nativeAdView = nil;
+    }
+    
+    self.nativeAdView = [[ANNativeRenderingViewController alloc] initWithSize:sizeofWebView BaseObject:nativeRenderingElement];
+     self.nativeAdView.loadingDelegate = self;
+}
+
+
+- (void) didFailToLoadNativeWebViewController{
+    if ([self.adObjectHandler isKindOfClass:[ANNativeStandardAdResponse class]]) {
+        ANNativeStandardAdResponse *nativeStandardAdResponse = (ANNativeStandardAdResponse *)self.adObjectHandler;
+        [self traditionalNativeAd:nativeStandardAdResponse];
+    }else{
+        NSError  *error  = ANError(@"ANAdWebViewController is UNDEFINED.", ANAdResponseInternalError);
+        ANAdFetcherResponse  *fetcherResponse = [ANAdFetcherResponse responseWithError:error];
+        [self processFinalResponse:fetcherResponse];
+    }
+}
+
+- (void) didCompleteFirstLoadFromNativeWebViewController:(ANNativeRenderingViewController *)controller{
+    ANAdFetcherResponse  *fetcherResponse  = nil;
+
+    if (self.nativeAdView == controller)
+    {
+        fetcherResponse = [ANAdFetcherResponse responseWithAdObject:controller andAdObjectHandler:self.adObjectHandler];
+        [self processFinalResponse:fetcherResponse];
+    } else {
+        [self didFailToLoadNativeWebViewController];
+    }
+}
 
 
 #pragma mark - ANUniversalAdFetcherDelegate.